Move and rename loadProjectConfig file 🤠
